+++ src/main/java/com/art/repository/PasswordResetTokenRepository.java
@@ -4,11 +4,9 @@ import com.art.model.PasswordResetToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
 public interface PasswordResetTokenRepository extends JpaRepository<PasswordResetToken, String> {
     PasswordResetToken findByToken(String token);
-    PasswordResetToken findById(BigInteger id);
-    void deleteById(BigInteger id);
+    PasswordResetToken findById(Long id);
+    void deleteById(Long id);
 }
