+++ src/main/java/com/art/repository/UserAnnexRepository.java
@@ -4,8 +4,6 @@ import com.art.model.UsersAnnexToContracts;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 /**
  * @author Alexandr Stegnin
  */
@@ -13,7 +11,7 @@ import java.math.BigInteger;
 @Repository
 public interface UserAnnexRepository extends JpaRepository<UsersAnnexToContracts, Long> {
 
-    Boolean existsByUserIdAndAnnexReadIs(BigInteger userId, Integer read);
+    Boolean existsByUserIdAndAnnexReadIs(Long userId, Integer read);
 
-    Boolean existsByUserIdAndDateReadIsNull(BigInteger userId);
+    Boolean existsByUserIdAndDateReadIsNull(Long userId);
 }
